requirements.txt
cryptography==41.0.3
pikepdf==8.7.0
Pillow>=10.0.1
python-docx==0.8.11
click==8.1.7
PyQt5==5.15.9
pyyaml==6.0
pytest==7.4.0
pytest-cov==4.1.0
